refactor: modular project structure and architecture documentation. Closes #9
