--- src/main/java/com/raullopezpenalva/newsletter_service/model/SubscriptionStatus.java
@@ -1,7 +1,0 @@
-package com.raullopezpenalva.newsletter_service.model;
-
-public enum SubscriptionStatus {
-    PENDING,
-    ACTIVE,
-    UNSUBSCRIBED
-}
