--- src/main/java/com/raullopezpenalva/newsletter_service/model/TokenType.java
@@ -1,6 +1,0 @@
-package com.raullopezpenalva.newsletter_service.model;
-
-public enum TokenType {
-    CONFIRMATION,
-    UNSUBSCRIBE
-}
