+++ src/main/java/com/raullopezpenalva/newsletter_service/modules/newsletter/domain/model/SubscriptionStatus.java
@@ -1,0 +1,7 @@
+package com.raullopezpenalva.newsletter_service.modules.newsletter.domain.model;
+
+public enum SubscriptionStatus {
+    PENDING,
+    ACTIVE,
+    UNSUBSCRIBED
+}
