+++ src/main/java/com/raullopezpenalva/newsletter_service/modules/newsletter/infrastructure/security/JwtService.java.txt
@@ -1,0 +1,107 @@
+package com.raullopezpenalva.newsletter_service.security;
+
+import io.jsonwebtoken.*;
+import io.jsonwebtoken.security.Keys;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Service;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.GrantedAuthority;
+
+import java.nio.charset.StandardCharsets;
+import java.security.Key;
+import java.util.Date;
+import java.util.stream.Collectors;
+
+@Service
+public class JwtService {
+
+    // JWT configuration properties
+    @Value("${app.security.jwt.secret}")
+    private String jwtSecret;
+
+    // Token validity in milliseconds (e.g., 3600000 ms = 1 hour)
+    @Value("${app.security.jwt.expirationMs}")
+    private int jwtExpirationMs;
+
+    /**
+     * Generates a cryptographic key for signing and verifying JWT tokens using the HMAC SHA algorithm.
+     * <p>
+     * The key is derived from the configured JWT secret string, encoded in UTF-8.
+     * This method ensures that the secret is properly converted to a {@link Key} object
+     * compatible with JWT operations.
+     *
+     * @return a {@link Key} instance for HMAC SHA signing and verification of JWTs
+     */
+    private Key key() {
+        return Keys.hmacShaKeyFor(jwtSecret.getBytes(StandardCharsets.UTF_8));
+    }
+
+    /**
+     * Generates a JWT token for the given user details.
+     *
+     * @param userDetails the user details to include in the token
+     * @return the generated JWT token as a string
+     */
+    public String generateToken(UserDetails userDetails) {
+        String authorities = userDetails.getAuthorities().stream()
+                .map(GrantedAuthority::getAuthority)
+                .collect(Collectors.joining(","));
+        Date now = new Date();
+        Date expiryDate = new Date(now.getTime() + jwtExpirationMs);
+
+        return Jwts.builder()
+                .setSubject(userDetails.getUsername())
+                .claim("roles", authorities)
+                .setIssuedAt(now)
+                .setExpiration(expiryDate)
+                .signWith(key(), SignatureAlgorithm.HS256)
+                .compact();
+    }
+
+    /**
+     * Extracts the username (subject) from the given JWT token.
+     *
+     * @param token the JWT token
+     * @return the username (subject) contained in the token
+     */
+    public String extractUsername(String token) {
+        return parseClaims(token).getBody().getSubject();
+    }
+
+    /**
+     * Validates the given JWT token against the provided user details.
+     *
+     * @param token       the JWT token to validate
+     * @param userDetails the user details to compare against
+     * @return true if the token is valid and matches the user details, false otherwise
+     */
+    public boolean isTokenValid(String token, UserDetails userDetails) {
+        final String username = extractUsername(token);
+        return (username.equals(userDetails.getUsername()) && !isTokenExpired(token));
+    }
+
+    /**
+     * Checks if the given JWT token has expired.
+     *
+     * @param token the JWT token to check
+     * @return true if the token has expired, false otherwise
+     */
+    private boolean isTokenExpired(String token) {
+        return parseClaims(token).getBody().getExpiration().before(new Date());
+    }
+
+    /**
+     * Parses the claims from the given JWT token.
+     *
+     * @param token the JWT token to parse
+     * @return the parsed claims as a {@link Jws<Claims>} object
+     * @throws JwtException if the token is invalid or cannot be parsed
+     */
+    private Jws<Claims> parseClaims(String token) {
+        return Jwts.parserBuilder()
+                .setSigningKey(key())
+                .build()
+                .parseClaimsJws(token);
+    }
+    
+}
