+++ src/main/java/com/raullopezpenalva/newsletter_service/modules/platform/tokens/domain/TokenType.java
@@ -1,0 +1,6 @@
+package com.raullopezpenalva.newsletter_service.modules.platform.tokens.domain;
+
+public enum TokenType {
+    CONFIRMATION,
+    UNSUBSCRIBE
+}
